requirements.txt
# Core dependencies
numpy>=1.20.0
opencv-python>=4.5.5
torch>=1.10.0
torchvision>=0.11.0
requests>=2.27.1
tqdm>=4.62.0
matplotlib>=3.5.0
pillow>=9.0.0

# Detection and tracking
ultralytics # For YOLOv8/v12
supervision>=0.25.1 # For ByteTrack, visualization
filterpy>=1.4.5     # For Kalman filter

# Re-identification
faiss-gpu>=1.7.0    # Vector database (use faiss-cpu for CPU-only systems)

# Audio processing
TTS>=0.10.0         # Text-to-speech
pygame>=2.1.0       # For audio playback

# Utilities
pyyaml>=6.0
rich>=12.0.0        # Pretty console output